organize favorite folder route function and querey in it's file relates #55
