+++ server/__test__/routes/getFavoriteCourses.test.js
@@ -54,7 +54,7 @@ describe('get request to /favorite', () => {
     ]);
   });
 
-  it('if user id does not exits will return Sign-in first', async () => {
+  it('if user id does not exist will return Sign-in first', async () => {
     expect.assertions(1);
     const { body } = await request(app)
       .get('/api/v1/favorite/3')
