+++ server/src/controllers/favoriteFolder/getFavoriteFolder.js
