+++ server/src/controllers/favoriteFolder/updateCourseToFolder.js
@@ -1,19 +1,19 @@
 const {
-  editFavoriteFolderQuery,
+  addCourseToFolderQuery,
   checkUserFolder,
-} = require('../../../database/queries');
-const { folderTitleSchema } = require('../../../utils');
+} = require('../../database/queries');
+const { addCourseToFolder } = require('../../utils');
 
 module.exports = async (req, res, next) => {
   const { id } = req.user;
-  const { title, folderId } = req.body;
+  const { courseId, folderId } = req.body;
   try {
+    await addCourseToFolder.validate({ courseId, folderId });
     const { rows } = await checkUserFolder(id);
     const isUserHaveFolder = rows.some((folder) => folder.id === folderId);
     if (isUserHaveFolder) {
-      await folderTitleSchema.validate({ title });
-      await editFavoriteFolderQuery(title, folderId);
-      res.json({ message: 'folder title updated successfully' });
+      await addCourseToFolderQuery(id, courseId, folderId);
+      res.json({ message: 'course assigned to folder successfully' });
     } else {
       res.status(401).json({ message: 'Un-Authorized' });
     }
