+++ server/src/database/queries/favoriteFolder/getFavoriteFolderQuery.js
@@ -1,4 +1,4 @@
-const connection = require('../config/connection');
+const connection = require('../../config/connection');
 
 const getFavoriteFolderQuery = (userId) =>
   connection.query('SELECT * FROM user_favorite_folders WHERE user_id = $1;', [
